requirements.txt
openai
vllm
fastapi
uvicorn
httpx
pandas
numpy
langgraph
faiss-cpu
python-dotenv
sentence-transformers